+++ include/dxc/DXIL/DxilConstants.h
@@ -15,6 +15,12 @@
 
 namespace hlsl {
 
+enum class UnusedResourceBinding : uint32_t { Strip, Reserved, KeepAll, Count };
+
+static_assert(UnusedResourceBinding::Count <= UnusedResourceBinding(7),
+              "Only 3 bits are reserved for UnusedResourceBinding by HLOptions "
+              "and ShaderFlags");
+
 /* <py>
 import hctdb_instrhelp
 </py> */