+++ tools/clang/lib/CodeGen/CGHLSLMS.cpp
@@ -400,6 +400,8 @@ CGMSHLSLRuntime::CGMSHLSLRuntime(CodeGenModule &CGM)
   opts.PackingStrategy = CGM.getCodeGenOpts().HLSLSignaturePackingStrategy;
   opts.bLegacyResourceReservation =
       CGM.getCodeGenOpts().HLSLLegacyResourceReservation;
+  opts.bUnusedResourceBinding =
+      unsigned(CGM.getCodeGenOpts().HLSLUnusedResourceBinding);
   opts.bForceZeroStoreLifetimes =
       CGM.getCodeGenOpts().HLSLForceZeroStoreLifetimes;